+++ NEWS.md
@@ -1,3 +1,8 @@
+# GPEDM 0.0.0.9003
+
+* Added `getrhomatrix` function, and `rhomatrix` argument to `fitGP`. This allows you to specify a fixed pairwise rho matrix for a hierarchical model.
+* Added vignette for VS-EDM and new example dataset.
+
 # GPEDM 0.0.0.9002
 
 * First logged development version.